+++ je_auto_control/utils/http_client/__init__.py
@@ -1,6 +1,6 @@
 """Dependency-free HTTP(S) client for AutoControl action steps."""
 from je_auto_control.utils.http_client.http_client import (
-    build_call, http_request, urllib_transport,
+    build_call, http_request, perform_call, urllib_transport,
 )
 
-__all__ = ["build_call", "http_request", "urllib_transport"]
+__all__ = ["build_call", "http_request", "perform_call", "urllib_transport"]